Balance type refactoring (#386)
