+++ engine-precompiles/src/lib.rs
@@ -52,7 +52,7 @@ impl From<PrecompileOutput> for evm::executor::PrecompileOutput {
     fn from(output: PrecompileOutput) -> Self {
         evm::executor::PrecompileOutput {
             exit_status: ExitSucceed::Returned,
-            cost: output.cost.into_u64(),
+            cost: output.cost.as_u64(),
             output: output.output,
             logs: output.logs,
         }
