+++ engine-sdk/src/near_runtime.rs
@@ -40,7 +40,7 @@ impl Runtime {
                 b"state_migration",
                 &[],
                 0,
-                Self::GAS_FOR_STATE_MIGRATION.into_u64(),
+                Self::GAS_FOR_STATE_MIGRATION.as_u64(),
             )
         }
     }
@@ -266,8 +266,8 @@ impl crate::promise::PromiseHandler for Runtime {
         let account_id = args.target_account_id.as_bytes();
         let method_name = args.method.as_bytes();
         let arguments = args.args.as_slice();
-        let amount = args.attached_balance;
-        let gas = args.attached_gas;
+        let amount = args.attached_balance.as_u128();
+        let gas = args.attached_gas.as_u64();
 
         let id = unsafe {
             exports::promise_create(
@@ -292,8 +292,8 @@ impl crate::promise::PromiseHandler for Runtime {
         let account_id = callback.target_account_id.as_bytes();
         let method_name = callback.method.as_bytes();
         let arguments = callback.args.as_slice();
-        let amount = callback.attached_balance;
-        let gas = callback.attached_gas;
+        let amount = callback.attached_balance.as_u128();
+        let gas = callback.attached_gas.as_u64();
 
         let id = unsafe {
             exports::promise_then(
@@ -322,7 +322,7 @@ impl crate::promise::PromiseHandler for Runtime {
         for action in args.actions.iter() {
             match action {
                 PromiseAction::Transfer { amount } => unsafe {
-                    let amount = *amount;
+                    let amount = amount.as_u128();
                     exports::promise_batch_action_transfer(id, &amount as *const u128 as _);
                 },
                 PromiseAction::DeployConotract { code } => unsafe {
@@ -341,7 +341,7 @@ impl crate::promise::PromiseHandler for Runtime {
                 } => unsafe {
                     let method_name = name.as_bytes();
                     let arguments = args.as_slice();
-                    let amount = *attached_yocto;
+                    let amount = attached_yocto.as_u128();
                     exports::promise_batch_action_function_call(
                         id,
                         method_name.len() as _,
@@ -349,7 +349,7 @@ impl crate::promise::PromiseHandler for Runtime {
                         arguments.len() as _,
                         arguments.as_ptr() as _,
                         &amount as *const u128 as _,
-                        *gas,
+                        gas.as_u64(),
                     )
                 },
             }
