+++ engine-tests/src/test_utils/mod.rs
@@ -1,4 +1,5 @@
 use aurora_engine_types::account_id::AccountId;
+use aurora_engine_types::types::NEP141Wei;
 use borsh::{BorshDeserialize, BorshSerialize};
 use near_primitives_core::config::VMConfig;
 use near_primitives_core::contract::ContractCode;
@@ -270,7 +271,8 @@ impl AuroraRunner {
                 .get(&ft_key)
                 .map(|bytes| FungibleToken::try_from_slice(&bytes).unwrap())
                 .unwrap_or_default();
-            current_ft.total_eth_supply_on_near += init_balance.raw().as_u128();
+            current_ft.total_eth_supply_on_near =
+                current_ft.total_eth_supply_on_near + NEP141Wei::new(init_balance.raw().as_u128());
             current_ft
         };
