+++ engine/src/connector.rs
@@ -9,9 +9,9 @@ use crate::parameters::{
     StorageWithdrawCallArgs, TransferCallArgs, TransferCallCallArgs, WithdrawResult,
 };
 use crate::prelude::{
-    format, sdk, str, validate_eth_address, AccountId, Address, Balance, BorshDeserialize,
-    BorshSerialize, EthAddress, EthConnectorStorageId, KeyPrefix, NearGas, PromiseResult, ToString,
-    Vec, WithdrawCallArgs, ERR_FAILED_PARSE, H160,
+    format, sdk, str, validate_eth_address, AccountId, Address, BorshDeserialize, BorshSerialize,
+    EthAddress, EthConnectorStorageId, KeyPrefix, NearGas, PromiseResult, ToString, Vec,
+    WithdrawCallArgs, Yocto, ERR_FAILED_PARSE, H160,
 };
 use crate::prelude::{
     AddressValidationError, PromiseBatchAction, PromiseCreateArgs, PromiseWithCallbackArgs,
@@ -19,10 +19,12 @@ use crate::prelude::{
 use crate::proof::Proof;
 use aurora_engine_sdk::env::Env;
 use aurora_engine_sdk::io::{StorageIntermediate, IO};
+use aurora_engine_types::types::{NEP141Wei, Wei, ZERO_NEP141_WEI, ZERO_WEI};
+use primitive_types::U256;
 
 pub const ERR_NOT_ENOUGH_BALANCE_FOR_FEE: &str = "ERR_NOT_ENOUGH_BALANCE_FOR_FEE";
 /// Indicate zero attached balance for promise call
-pub const ZERO_ATTACHED_BALANCE: Balance = 0;
+pub const ZERO_ATTACHED_BALANCE: Yocto = Yocto::new(0);
 /// NEAR Gas for calling `fininsh_deposit` promise. Used in the `deposit` logic.
 pub const GAS_FOR_FINISH_DEPOSIT: NearGas = NearGas::new(50_000_000_000_000);
 /// NEAR Gas for calling `verify_log_entry` promise. Used in the `deposit` logic.
@@ -160,7 +162,7 @@ impl<I: IO + Copy> EthConnectorContract<I> {
             return Err(error::DepositError::CustodianAddressMismatch);
         }
 
-        if event.fee.into_u128() >= event.amount {
+        if NEP141Wei::new(event.fee.as_u128()) >= event.amount {
             return Err(error::DepositError::InsufficientAmountForFee);
         }
 
@@ -180,7 +182,7 @@ impl<I: IO + Copy> EthConnectorContract<I> {
             method: "verify_log_entry".to_string(),
             args: proof_to_verify,
             attached_balance: ZERO_ATTACHED_BALANCE,
-            attached_gas: GAS_FOR_VERIFY_LOG_ENTRY.into_u64(),
+            attached_gas: GAS_FOR_VERIFY_LOG_ENTRY,
         };
 
         // Finalize deposit
@@ -232,7 +234,7 @@ impl<I: IO + Copy> EthConnectorContract<I> {
             method: "finish_deposit".to_string(),
             args: data,
             attached_balance: ZERO_ATTACHED_BALANCE,
-            attached_gas: GAS_FOR_FINISH_DEPOSIT.into_u64(),
+            attached_gas: GAS_FOR_FINISH_DEPOSIT,
         };
         Ok(PromiseWithCallbackArgs {
             base: verify_call,
@@ -274,9 +276,9 @@ impl<I: IO + Copy> EthConnectorContract<I> {
             // Mint - calculate new balances
             self.mint_eth_on_near(
                 data.new_owner_id.clone(),
-                data.amount - data.fee.into_u128(),
+                data.amount - NEP141Wei::new(data.fee.as_u128()),
             )?;
-            self.mint_eth_on_near(data.relayer_id, data.fee.into_u128())?;
+            self.mint_eth_on_near(data.relayer_id, NEP141Wei::new(data.fee.as_u128()))?;
             // Store proof only after `mint` calculations
             self.record_proof(&data.proof_key)?;
             // Save new contract data
@@ -289,7 +291,7 @@ impl<I: IO + Copy> EthConnectorContract<I> {
     pub(crate) fn internal_remove_eth(
         &mut self,
         address: &Address,
-        amount: Balance,
+        amount: Wei,
     ) -> Result<(), fungible_token::error::WithdrawError> {
         self.burn_eth_on_aurora(address.0, amount)?;
         self.save_ft_contract();
@@ -312,12 +314,12 @@ impl<I: IO + Copy> EthConnectorContract<I> {
     fn mint_eth_on_near(
         &mut self,
         owner_id: AccountId,
-        amount: Balance,
+        amount: NEP141Wei,
     ) -> Result<(), fungible_token::error::DepositError> {
         sdk::log!(&format!("Mint {} nETH tokens for: {}", amount, owner_id));
 
         if self.ft.get_account_eth_balance(&owner_id).is_none() {
-            self.ft.accounts_insert(&owner_id, 0);
+            self.ft.accounts_insert(&owner_id, ZERO_NEP141_WEI);
         }
         self.ft.internal_deposit_eth_to_near(&owner_id, amount)
     }
@@ -326,7 +328,7 @@ impl<I: IO + Copy> EthConnectorContract<I> {
     fn mint_eth_on_aurora(
         &mut self,
         owner_id: EthAddress,
-        amount: Balance,
+        amount: Wei,
     ) -> Result<(), fungible_token::error::DepositError> {
         sdk::log!(&format!(
             "Mint {} ETH tokens for: {}",
@@ -340,7 +342,7 @@ impl<I: IO + Copy> EthConnectorContract<I> {
     fn burn_eth_on_aurora(
         &mut self,
         address: EthAddress,
-        amount: Balance,
+        amount: Wei,
     ) -> Result<(), fungible_token::error::WithdrawError> {
         sdk::log!(&format!(
             "Burn {} ETH tokens for: {}",
@@ -412,7 +414,7 @@ impl<I: IO + Copy> EthConnectorContract<I> {
     ) -> Result<(), crate::prelude::types::error::BalanceOverflowError> {
         let balance = self
             .ft
-            .internal_unwrap_balance_of_eth_on_aurora(args.address)?;
+            .internal_unwrap_balance_of_eth_on_aurora(args.address);
         sdk::log!(&format!(
             "Balance of ETH [{}]: {}",
             hex::encode(args.address),
@@ -487,7 +489,7 @@ impl<I: IO + Copy> EthConnectorContract<I> {
             let message_data = FtTransferMessageData::parse_on_transfer_message(&args.msg)
                 .map_err(error::FtTransferCallError::MessageParseFailed)?;
             // Check is transfer amount > fee
-            if message_data.fee.into_u128() >= args.amount {
+            if message_data.fee.as_u128() >= args.amount.as_u128() {
                 return Err(error::FtTransferCallError::InsufficientAmountForFee);
             }
 
@@ -496,9 +498,11 @@ impl<I: IO + Copy> EthConnectorContract<I> {
             // Note: It can't overflow because the total supply doesn't change during transfer.
             let amount_for_check = self
                 .ft
-                .internal_unwrap_balance_of_eth_on_aurora(message_data.recipient)
-                .map_err(error::FtTransferCallError::BalanceOverflow)?;
-            if amount_for_check.checked_add(args.amount).is_none() {
+                .internal_unwrap_balance_of_eth_on_aurora(message_data.recipient);
+            if amount_for_check
+                .checked_add(Wei::from(args.amount))
+                .is_none()
+            {
                 return Err(error::FtTransferCallError::Transfer(
                     fungible_token::error::TransferError::BalanceOverflow,
                 ));
@@ -506,7 +510,7 @@ impl<I: IO + Copy> EthConnectorContract<I> {
             if self
                 .ft
                 .total_eth_supply_on_aurora
-                .checked_add(args.amount)
+                .checked_add(Wei::from(args.amount))
                 .is_none()
             {
                 return Err(error::FtTransferCallError::Transfer(
@@ -532,7 +536,7 @@ impl<I: IO + Copy> EthConnectorContract<I> {
     pub fn storage_deposit(
         &mut self,
         predecessor_account_id: AccountId,
-        amount: Balance,
+        amount: Yocto,
         args: StorageDepositCallArgs,
     ) -> Result<Option<PromiseBatchAction>, fungible_token::error::StorageFundingError> {
         let account_id = args
@@ -599,15 +603,21 @@ impl<I: IO + Copy> EthConnectorContract<I> {
             .map_err(error::FtTransferCallError::MessageParseFailed)?;
 
         // Special case when predecessor_account_id is current_account_id
-        let fee = message_data.fee.into_u128();
+        let wei_fee = Wei::from(message_data.fee);
         // Mint fee to relayer
         let relayer = engine.get_relayer(message_data.relayer.as_bytes());
-        match (fee, relayer) {
-            (fee, Some(H160(evm_relayer_address))) if fee > 0 => {
-                self.mint_eth_on_aurora(message_data.recipient, args.amount - fee)?;
+        match (wei_fee, relayer) {
+            (fee, Some(H160(evm_relayer_address))) if fee > ZERO_WEI => {
+                self.mint_eth_on_aurora(
+                    message_data.recipient,
+                    Wei::new(U256::from(args.amount.as_u128())) - fee,
+                )?;
                 self.mint_eth_on_aurora(evm_relayer_address, fee)?;
             }
-            _ => self.mint_eth_on_aurora(message_data.recipient, args.amount)?,
+            _ => self.mint_eth_on_aurora(
+                message_data.recipient,
+                Wei::new(U256::from(args.amount.as_u128())),
+            )?,
         }
         self.save_ft_contract();
         self.io.return_output("\"0\"".as_bytes());
