+++ engine/src/transaction/eip_2930.rs
@@ -1,5 +1,6 @@
 use crate::prelude::precompiles::secp256k1::ecrecover;
-use crate::prelude::{sdk, Address, Vec, Wei, H256, U256};
+use crate::prelude::Wei;
+use crate::prelude::{sdk, Address, Vec, H256, U256};
 use rlp::{Decodable, DecoderError, Encodable, Rlp, RlpStream};
 
 /// Type indicator (per EIP-2718) for access list transactions
